+++ BergerMsfaApi/Models/DemandGeneration/LeadFollowUpModel.cs
@@ -40,7 +40,7 @@ namespace BergerMsfaApi.Models.DemandGeneration
         public string PaintContractorNameChangeReason { get; set; }
         public string PaintContractorMobile { get; set; }
         public string PaintContractorMobileChangeReason { get; set; }
-        public int NumberOfStoriedBuilding { get; set; }
+        public double NumberOfStoriedBuilding { get; set; }
         public string NumberOfStoriedBuildingChangeReason { get; set; }
         public decimal ExpectedValue { get; set; }
         //public int ExpectedValueChangeCount { get; set; }
@@ -210,7 +210,7 @@ namespace BergerMsfaApi.Models.DemandGeneration
         public string PaintContractorNameChangeReason { get; set; }
         public string PaintContractorMobile { get; set; }
         public string PaintContractorMobileChangeReason { get; set; }
-        public int NumberOfStoriedBuilding { get; set; }
+        public double NumberOfStoriedBuilding { get; set; }
         public string NumberOfStoriedBuildingChangeReason { get; set; }
         public decimal ExpectedValue { get; set; }
         public int ExpectedValueChangeCount { get; set; }
